Fix YouTube api not always returning all subscriptions
It seems that when `order` is `relevance`,
it does not return all subscriptions. I have seen it return the same
subscription multiple times. Maybe the definition of "most relevant"
changes over time, returning the same thing again in place of something
that is never returned.

The most reliable ordering is therefore alphabetical.