Don't use data folder for docker method
Use a few more volumes in the docker-compose.yml instead. This way is
more consistent between both methods. It may also be easier for users
new to Docker to change where a particular file is stored. For example,
I personally modify the volume for the database to be in a different
directory so I can keep my docker-compose and tokens in version control
and keep my database on my zfs pool.